Remove a few group columns from editing interface
Removed is_use_landing and is_use_captcha, which I don't intend to support in
the first release. Also removed system_group_type, as I am not sure how this
should be edited. This gets a value from one of the DownloadGroupPeer consts
according to whether the group should be the default for (1) a file with no
group or (2) a newly created group. I'll review this later.